♻️ Refactor: signUp페이지의 형태를 일반적인 container구조로 변경 및 일부 에러 제거
