+++ pages/signup.tsx
@@ -1,15 +1,95 @@
+import signBlurError from "@/src/utils/sign-blur-error-message/signBlurError";
+import { FormEvent, RefObject, useRef, useState } from "react";
 import router from "next/router";
-import * as S from "@/styles/pages/sign-page.style";
+import axios from "axios";
+import { SIGN_INPUT_ERROR_MESSAGES } from "@/src/constant/SIGN_INPUT_TEXTS";
+import emailBlur from "@/src/utils/sign-blur-error-message/input-blur-types/emailBlur";
+import passwordBlur from "@/src/utils/sign-blur-error-message/input-blur-types/passwordBlur";
+import passCheckBlur from "@/src/utils/sign-blur-error-message/input-blur-types/passCheckBlur";
+import SignUpPresenter from "@/src/containers/signup-page/signup.presenter";
 
-// Components
-import SignInPageContainer from "@/src/containers/signup-pagea/useSignUpPage";
-import SignPageHeader from "@/src/components/layout/sign-page/header/SignPageHeader";
-import SignInput from "@/src/components/sign-input/SignInput";
-import SocialSign from "@/src/components/layout/sign-page/social/SocialSign";
-import { useEffect } from "react";
+/**
+ * @description signupPage의 전반적인 로직이 담겨있습니다.
+ * @returns
+ */
 
 export default function SignUp() {
-  const {
+  const [emailError, setEmailError] = useState("");
+  const [passwordError, setPasswordError] = useState("");
+  const [passCheckError, setPassCheckError] = useState("");
+  const emailRef = useRef<HTMLInputElement>(null);
+  const passwordRef = useRef<HTMLInputElement>(null);
+  const passCheckRef = useRef<HTMLInputElement>(null);
+
+  const blurEvent = (type: string, ref: RefObject<HTMLInputElement>) => {
+    const userInput = ref.current!.value;
+    const comparePassword =
+      type === "password-check" ? passwordRef.current!.value : undefined;
+    let setFunction;
+
+    switch (type) {
+      case "email":
+        setFunction = setEmailError;
+        break;
+
+      case "password-sign-up":
+        setFunction = setPasswordError;
+        break;
+
+      case "password-check":
+        setFunction = setPassCheckError;
+        break;
+
+      default:
+        console.error("input의 타입 지정이 올바르지 않음");
+        return;
+    }
+    signBlurError(userInput, type, setFunction, comparePassword);
+  };
+
+  const submitEvent = async (e: FormEvent) => {
+    e.preventDefault();
+    const emailInput = emailRef.current!.value;
+    const passwordInput = passwordRef.current!.value;
+    const passCheckInput = passCheckRef.current!.value;
+    setEmailError("");
+    setPasswordError("");
+    setPassCheckError("");
+
+    try {
+      await axios.post("https://bootcamp-api.codeit.kr/api/check-email", {
+        email: emailInput,
+      });
+
+      if (
+        emailBlur(emailInput) ||
+        passwordBlur(passwordInput, "password-sign-up") ||
+        passCheckBlur(passCheckInput, passwordInput)
+      ) {
+        throw new Error("유효하지 않은 회원가입 시도");
+      }
+
+      const { data }: { data: { accessToken: string; refreshToken: string } } =
+        await axios.post("https://bootcamp-api.codeit.kr/api/sign-up", {
+          email: emailInput,
+          password: passwordInput,
+        });
+      localStorage.setItem("accessToken", data.accessToken);
+      router.push("/folder");
+    } catch (e: any) {
+      setEmailError(SIGN_INPUT_ERROR_MESSAGES.NOT_CORRECT_EMAIL);
+      setPasswordError(SIGN_INPUT_ERROR_MESSAGES.NOT_CORRECT_PASSWORD);
+
+      signBlurError(emailInput, "email", setEmailError);
+      signBlurError(passwordInput, "password-sign-up", setPasswordError);
+      signBlurError(passCheckInput, "password-check", setPassCheckError);
+      if (e.response.status === 409) {
+        setEmailError(SIGN_INPUT_ERROR_MESSAGES.DUPLICATE_EMAIL);
+      }
+    }
+  };
+
+  const props = {
     emailError,
     passwordError,
     passCheckError,
@@ -18,45 +98,7 @@ export default function SignUp() {
     passCheckRef,
     blurEvent,
     submitEvent,
-  } = SignInPageContainer();
+  };
 
-  useEffect(() => {
-    if (localStorage.getItem("accessToken")) {
-      router.push("/folder");
-    }
-  });
-
-  return (
-    <>
-      <S.SignPageWrapper>
-        <SignPageHeader
-          text="이미 회원이신가요?"
-          linkText="로그인 하기"
-          linkTo="/signin"
-        />
-        <S.SignFormWrapper id="sign" onSubmit={(e) => submitEvent(e)}>
-          <SignInput
-            inputRef={emailRef}
-            inputType="email"
-            errorMessage={emailError}
-            blurEvent={() => blurEvent("email", emailRef)}
-          />
-          <SignInput
-            inputRef={passwordRef}
-            inputType="password-sign-up"
-            errorMessage={passwordError}
-            blurEvent={() => blurEvent("password-sign-up", passwordRef)}
-          />
-          <SignInput
-            inputRef={passCheckRef}
-            inputType="password-check"
-            errorMessage={passCheckError}
-            blurEvent={() => blurEvent("password-check", passCheckRef)}
-          />
-        </S.SignFormWrapper>
-        <S.SignSubmitButton form="sign">회원가입</S.SignSubmitButton>
-        <SocialSign />
-      </S.SignPageWrapper>
-    </>
-  );
+  return <SignUpPresenter props={props} />;
 }
